+++ client/test/api_cron_job_service_test.go
@@ -253,6 +253,17 @@ func Test_taikuncore_CronJobServiceAPIService(t *testing.T) {
 
 	})
 
+	t.Run("Test CronJobServiceAPIService CronjobSyncDnsCertConfigs", func(t *testing.T) {
+
+		t.Skip("skip test")  // remove to run test
+
+		httpRes, err := apiClient.CronJobServiceAPI.CronjobSyncDnsCertConfigs(context.Background()).Execute()
+
+		require.Nil(t, err)
+		assert.Equal(t, 200, httpRes.StatusCode)
+
+	})
+
 	t.Run("Test CronJobServiceAPIService CronjobSyncOpaProfiles", func(t *testing.T) {
 
 		t.Skip("skip test")  // remove to run test